+++ sdk/java/talon-client/src/main/java/talon/data/Data.java
@@ -4446,2688 +4446,6 @@ java.lang.String defaultValue) {
 
   }
 
-  public interface FileUploadOrBuilder extends
-      // @@protoc_insertion_point(interface_extends:talon.data.FileUpload)
-      com.google.protobuf.MessageOrBuilder {
-
-    /**
-     * <code>string id = 1;</code>
-     * @return The id.
-     */
-    java.lang.String getId();
-    /**
-     * <code>string id = 1;</code>
-     * @return The bytes for id.
-     */
-    com.google.protobuf.ByteString
-        getIdBytes();
-
-    /**
-     * <code>string namespace = 2;</code>
-     * @return The namespace.
-     */
-    java.lang.String getNamespace();
-    /**
-     * <code>string namespace = 2;</code>
-     * @return The bytes for namespace.
-     */
-    com.google.protobuf.ByteString
-        getNamespaceBytes();
-
-    /**
-     * <code>string file_name = 3;</code>
-     * @return The fileName.
-     */
-    java.lang.String getFileName();
-    /**
-     * <code>string file_name = 3;</code>
-     * @return The bytes for fileName.
-     */
-    com.google.protobuf.ByteString
-        getFileNameBytes();
-
-    /**
-     * <code>string file_uid = 4;</code>
-     * @return The fileUid.
-     */
-    java.lang.String getFileUid();
-    /**
-     * <code>string file_uid = 4;</code>
-     * @return The bytes for fileUid.
-     */
-    com.google.protobuf.ByteString
-        getFileUidBytes();
-
-    /**
-     * <code>string path = 5;</code>
-     * @return The path.
-     */
-    java.lang.String getPath();
-    /**
-     * <code>string path = 5;</code>
-     * @return The bytes for path.
-     */
-    com.google.protobuf.ByteString
-        getPathBytes();
-
-    /**
-     * <code>string media_type = 6;</code>
-     * @return The mediaType.
-     */
-    java.lang.String getMediaType();
-    /**
-     * <code>string media_type = 6;</code>
-     * @return The bytes for mediaType.
-     */
-    com.google.protobuf.ByteString
-        getMediaTypeBytes();
-
-    /**
-     * <code>int32 purpose = 7;</code>
-     * @return The purpose.
-     */
-    int getPurpose();
-
-    /**
-     * <code>int32 index_policy = 8;</code>
-     * @return The indexPolicy.
-     */
-    int getIndexPolicy();
-
-    /**
-     * <code>int32 retention = 9;</code>
-     * @return The retention.
-     */
-    int getRetention();
-
-    /**
-     * <code>string object_key = 10;</code>
-     * @return The objectKey.
-     */
-    java.lang.String getObjectKey();
-    /**
-     * <code>string object_key = 10;</code>
-     * @return The bytes for objectKey.
-     */
-    com.google.protobuf.ByteString
-        getObjectKeyBytes();
-
-    /**
-     * <code>uint64 expected_size_bytes = 11;</code>
-     * @return The expectedSizeBytes.
-     */
-    long getExpectedSizeBytes();
-
-    /**
-     * <code>string expected_sha256 = 12;</code>
-     * @return The expectedSha256.
-     */
-    java.lang.String getExpectedSha256();
-    /**
-     * <code>string expected_sha256 = 12;</code>
-     * @return The bytes for expectedSha256.
-     */
-    com.google.protobuf.ByteString
-        getExpectedSha256Bytes();
-
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    int getRequiredHeadersCount();
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    boolean containsRequiredHeaders(
-        java.lang.String key);
-    /**
-     * Use {@link #getRequiredHeadersMap()} instead.
-     */
-    @java.lang.Deprecated
-    java.util.Map<java.lang.String, java.lang.String>
-    getRequiredHeaders();
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    java.util.Map<java.lang.String, java.lang.String>
-    getRequiredHeadersMap();
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    /* nullable */
-java.lang.String getRequiredHeadersOrDefault(
-        java.lang.String key,
-        /* nullable */
-java.lang.String defaultValue);
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    java.lang.String getRequiredHeadersOrThrow(
-        java.lang.String key);
-
-    /**
-     * <code>string created_by_agent = 14;</code>
-     * @return The createdByAgent.
-     */
-    java.lang.String getCreatedByAgent();
-    /**
-     * <code>string created_by_agent = 14;</code>
-     * @return The bytes for createdByAgent.
-     */
-    com.google.protobuf.ByteString
-        getCreatedByAgentBytes();
-
-    /**
-     * <code>string created_by_session_id = 15;</code>
-     * @return The createdBySessionId.
-     */
-    java.lang.String getCreatedBySessionId();
-    /**
-     * <code>string created_by_session_id = 15;</code>
-     * @return The bytes for createdBySessionId.
-     */
-    com.google.protobuf.ByteString
-        getCreatedBySessionIdBytes();
-
-    /**
-     * <code>int64 expires_at = 16;</code>
-     * @return The expiresAt.
-     */
-    long getExpiresAt();
-
-    /**
-     * <code>int64 created_at = 17;</code>
-     * @return The createdAt.
-     */
-    long getCreatedAt();
-
-    /**
-     * <code>bool consumed = 18;</code>
-     * @return The consumed.
-     */
-    boolean getConsumed();
-  }
-  /**
-   * <pre>
-   * Pending direct-to-object-storage upload for a namespace File.
-   *
-   * The target File is not committed to this object until CompleteFileUpload
-   * validates the uploaded object metadata and patches File.status.objectRef.
-   * </pre>
-   *
-   * Protobuf type {@code talon.data.FileUpload}
-   */
-  public static final class FileUpload extends
-      com.google.protobuf.GeneratedMessage implements
-      // @@protoc_insertion_point(message_implements:talon.data.FileUpload)
-      FileUploadOrBuilder {
-  private static final long serialVersionUID = 0L;
-    static {
-      com.google.protobuf.RuntimeVersion.validateProtobufGencodeVersion(
-        com.google.protobuf.RuntimeVersion.RuntimeDomain.PUBLIC,
-        /* major= */ 4,
-        /* minor= */ 34,
-        /* patch= */ 1,
-        /* suffix= */ "",
-        "FileUpload");
-    }
-    // Use FileUpload.newBuilder() to construct.
-    private FileUpload(com.google.protobuf.GeneratedMessage.Builder<?> builder) {
-      super(builder);
-    }
-    private FileUpload() {
-      id_ = "";
-      namespace_ = "";
-      fileName_ = "";
-      fileUid_ = "";
-      path_ = "";
-      mediaType_ = "";
-      objectKey_ = "";
-      expectedSha256_ = "";
-      createdByAgent_ = "";
-      createdBySessionId_ = "";
-    }
-
-    public static final com.google.protobuf.Descriptors.Descriptor
-        getDescriptor() {
-      return talon.data.Data.internal_static_talon_data_FileUpload_descriptor;
-    }
-
-    @java.lang.Override
-    public com.google.protobuf.Descriptors.Descriptor getDescriptorForType() {
-      return talon.data.Data.internal_static_talon_data_FileUpload_descriptor;
-    }
-
-    @SuppressWarnings({"rawtypes"})
-    @java.lang.Override
-    protected com.google.protobuf.MapFieldReflectionAccessor internalGetMapFieldReflection(
-        int number) {
-      switch (number) {
-        case 13:
-          return internalGetRequiredHeaders();
-        default:
-          throw new RuntimeException(
-              "Invalid map field number: " + number);
-      }
-    }
-    @java.lang.Override
-    protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-        internalGetFieldAccessorTable() {
-      return talon.data.Data.internal_static_talon_data_FileUpload_fieldAccessorTable
-          .ensureFieldAccessorsInitialized(
-              talon.data.Data.FileUpload.class, talon.data.Data.FileUpload.Builder.class);
-    }
-
-    public static final int ID_FIELD_NUMBER = 1;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object id_ = "";
-    /**
-     * <code>string id = 1;</code>
-     * @return The id.
-     */
-    @java.lang.Override
-    public java.lang.String getId() {
-      java.lang.Object ref = id_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        id_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string id = 1;</code>
-     * @return The bytes for id.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getIdBytes() {
-      java.lang.Object ref = id_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        id_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int NAMESPACE_FIELD_NUMBER = 2;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object namespace_ = "";
-    /**
-     * <code>string namespace = 2;</code>
-     * @return The namespace.
-     */
-    @java.lang.Override
-    public java.lang.String getNamespace() {
-      java.lang.Object ref = namespace_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        namespace_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string namespace = 2;</code>
-     * @return The bytes for namespace.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getNamespaceBytes() {
-      java.lang.Object ref = namespace_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        namespace_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int FILE_NAME_FIELD_NUMBER = 3;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object fileName_ = "";
-    /**
-     * <code>string file_name = 3;</code>
-     * @return The fileName.
-     */
-    @java.lang.Override
-    public java.lang.String getFileName() {
-      java.lang.Object ref = fileName_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        fileName_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string file_name = 3;</code>
-     * @return The bytes for fileName.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getFileNameBytes() {
-      java.lang.Object ref = fileName_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        fileName_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int FILE_UID_FIELD_NUMBER = 4;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object fileUid_ = "";
-    /**
-     * <code>string file_uid = 4;</code>
-     * @return The fileUid.
-     */
-    @java.lang.Override
-    public java.lang.String getFileUid() {
-      java.lang.Object ref = fileUid_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        fileUid_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string file_uid = 4;</code>
-     * @return The bytes for fileUid.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getFileUidBytes() {
-      java.lang.Object ref = fileUid_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        fileUid_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int PATH_FIELD_NUMBER = 5;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object path_ = "";
-    /**
-     * <code>string path = 5;</code>
-     * @return The path.
-     */
-    @java.lang.Override
-    public java.lang.String getPath() {
-      java.lang.Object ref = path_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        path_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string path = 5;</code>
-     * @return The bytes for path.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getPathBytes() {
-      java.lang.Object ref = path_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        path_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int MEDIA_TYPE_FIELD_NUMBER = 6;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object mediaType_ = "";
-    /**
-     * <code>string media_type = 6;</code>
-     * @return The mediaType.
-     */
-    @java.lang.Override
-    public java.lang.String getMediaType() {
-      java.lang.Object ref = mediaType_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        mediaType_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string media_type = 6;</code>
-     * @return The bytes for mediaType.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getMediaTypeBytes() {
-      java.lang.Object ref = mediaType_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        mediaType_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int PURPOSE_FIELD_NUMBER = 7;
-    private int purpose_ = 0;
-    /**
-     * <code>int32 purpose = 7;</code>
-     * @return The purpose.
-     */
-    @java.lang.Override
-    public int getPurpose() {
-      return purpose_;
-    }
-
-    public static final int INDEX_POLICY_FIELD_NUMBER = 8;
-    private int indexPolicy_ = 0;
-    /**
-     * <code>int32 index_policy = 8;</code>
-     * @return The indexPolicy.
-     */
-    @java.lang.Override
-    public int getIndexPolicy() {
-      return indexPolicy_;
-    }
-
-    public static final int RETENTION_FIELD_NUMBER = 9;
-    private int retention_ = 0;
-    /**
-     * <code>int32 retention = 9;</code>
-     * @return The retention.
-     */
-    @java.lang.Override
-    public int getRetention() {
-      return retention_;
-    }
-
-    public static final int OBJECT_KEY_FIELD_NUMBER = 10;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object objectKey_ = "";
-    /**
-     * <code>string object_key = 10;</code>
-     * @return The objectKey.
-     */
-    @java.lang.Override
-    public java.lang.String getObjectKey() {
-      java.lang.Object ref = objectKey_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        objectKey_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string object_key = 10;</code>
-     * @return The bytes for objectKey.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getObjectKeyBytes() {
-      java.lang.Object ref = objectKey_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        objectKey_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int EXPECTED_SIZE_BYTES_FIELD_NUMBER = 11;
-    private long expectedSizeBytes_ = 0L;
-    /**
-     * <code>uint64 expected_size_bytes = 11;</code>
-     * @return The expectedSizeBytes.
-     */
-    @java.lang.Override
-    public long getExpectedSizeBytes() {
-      return expectedSizeBytes_;
-    }
-
-    public static final int EXPECTED_SHA256_FIELD_NUMBER = 12;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object expectedSha256_ = "";
-    /**
-     * <code>string expected_sha256 = 12;</code>
-     * @return The expectedSha256.
-     */
-    @java.lang.Override
-    public java.lang.String getExpectedSha256() {
-      java.lang.Object ref = expectedSha256_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        expectedSha256_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string expected_sha256 = 12;</code>
-     * @return The bytes for expectedSha256.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getExpectedSha256Bytes() {
-      java.lang.Object ref = expectedSha256_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        expectedSha256_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int REQUIRED_HEADERS_FIELD_NUMBER = 13;
-    private static final class RequiredHeadersDefaultEntryHolder {
-      static final com.google.protobuf.MapEntry<
-          java.lang.String, java.lang.String> defaultEntry =
-              com.google.protobuf.MapEntry
-              .<java.lang.String, java.lang.String>newDefaultInstance(
-                  talon.data.Data.internal_static_talon_data_FileUpload_RequiredHeadersEntry_descriptor,
-                  com.google.protobuf.WireFormat.FieldType.STRING,
-                  "",
-                  com.google.protobuf.WireFormat.FieldType.STRING,
-                  "");
-    }
-    @SuppressWarnings("serial")
-    private com.google.protobuf.MapField<
-        java.lang.String, java.lang.String> requiredHeaders_;
-    private com.google.protobuf.MapField<java.lang.String, java.lang.String>
-    internalGetRequiredHeaders() {
-      if (requiredHeaders_ == null) {
-        return com.google.protobuf.MapField.emptyMapField(
-            RequiredHeadersDefaultEntryHolder.defaultEntry);
-      }
-      return requiredHeaders_;
-    }
-    public int getRequiredHeadersCount() {
-      return internalGetRequiredHeaders().getMap().size();
-    }
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    @java.lang.Override
-    public boolean containsRequiredHeaders(
-        java.lang.String key) {
-      if (key == null) { throw new NullPointerException("map key"); }
-      return internalGetRequiredHeaders().getMap().containsKey(key);
-    }
-    /**
-     * Use {@link #getRequiredHeadersMap()} instead.
-     */
-    @java.lang.Override
-    @java.lang.Deprecated
-    public java.util.Map<java.lang.String, java.lang.String> getRequiredHeaders() {
-      return getRequiredHeadersMap();
-    }
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    @java.lang.Override
-    public java.util.Map<java.lang.String, java.lang.String> getRequiredHeadersMap() {
-      return internalGetRequiredHeaders().getMap();
-    }
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    @java.lang.Override
-    public /* nullable */
-java.lang.String getRequiredHeadersOrDefault(
-        java.lang.String key,
-        /* nullable */
-java.lang.String defaultValue) {
-      if (key == null) { throw new NullPointerException("map key"); }
-      java.util.Map<java.lang.String, java.lang.String> map =
-          internalGetRequiredHeaders().getMap();
-      return map.containsKey(key) ? map.get(key) : defaultValue;
-    }
-    /**
-     * <code>map&lt;string, string&gt; required_headers = 13;</code>
-     */
-    @java.lang.Override
-    public java.lang.String getRequiredHeadersOrThrow(
-        java.lang.String key) {
-      if (key == null) { throw new NullPointerException("map key"); }
-      java.util.Map<java.lang.String, java.lang.String> map =
-          internalGetRequiredHeaders().getMap();
-      if (!map.containsKey(key)) {
-        throw new java.lang.IllegalArgumentException();
-      }
-      return map.get(key);
-    }
-
-    public static final int CREATED_BY_AGENT_FIELD_NUMBER = 14;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object createdByAgent_ = "";
-    /**
-     * <code>string created_by_agent = 14;</code>
-     * @return The createdByAgent.
-     */
-    @java.lang.Override
-    public java.lang.String getCreatedByAgent() {
-      java.lang.Object ref = createdByAgent_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        createdByAgent_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string created_by_agent = 14;</code>
-     * @return The bytes for createdByAgent.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getCreatedByAgentBytes() {
-      java.lang.Object ref = createdByAgent_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        createdByAgent_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int CREATED_BY_SESSION_ID_FIELD_NUMBER = 15;
-    @SuppressWarnings("serial")
-    private volatile java.lang.Object createdBySessionId_ = "";
-    /**
-     * <code>string created_by_session_id = 15;</code>
-     * @return The createdBySessionId.
-     */
-    @java.lang.Override
-    public java.lang.String getCreatedBySessionId() {
-      java.lang.Object ref = createdBySessionId_;
-      if (ref instanceof java.lang.String) {
-        return (java.lang.String) ref;
-      } else {
-        com.google.protobuf.ByteString bs =
-            (com.google.protobuf.ByteString) ref;
-        java.lang.String s = bs.toStringUtf8();
-        createdBySessionId_ = s;
-        return s;
-      }
-    }
-    /**
-     * <code>string created_by_session_id = 15;</code>
-     * @return The bytes for createdBySessionId.
-     */
-    @java.lang.Override
-    public com.google.protobuf.ByteString
-        getCreatedBySessionIdBytes() {
-      java.lang.Object ref = createdBySessionId_;
-      if (ref instanceof java.lang.String) {
-        com.google.protobuf.ByteString b =
-            com.google.protobuf.ByteString.copyFromUtf8(
-                (java.lang.String) ref);
-        createdBySessionId_ = b;
-        return b;
-      } else {
-        return (com.google.protobuf.ByteString) ref;
-      }
-    }
-
-    public static final int EXPIRES_AT_FIELD_NUMBER = 16;
-    private long expiresAt_ = 0L;
-    /**
-     * <code>int64 expires_at = 16;</code>
-     * @return The expiresAt.
-     */
-    @java.lang.Override
-    public long getExpiresAt() {
-      return expiresAt_;
-    }
-
-    public static final int CREATED_AT_FIELD_NUMBER = 17;
-    private long createdAt_ = 0L;
-    /**
-     * <code>int64 created_at = 17;</code>
-     * @return The createdAt.
-     */
-    @java.lang.Override
-    public long getCreatedAt() {
-      return createdAt_;
-    }
-
-    public static final int CONSUMED_FIELD_NUMBER = 18;
-    private boolean consumed_ = false;
-    /**
-     * <code>bool consumed = 18;</code>
-     * @return The consumed.
-     */
-    @java.lang.Override
-    public boolean getConsumed() {
-      return consumed_;
-    }
-
-    private byte memoizedIsInitialized = -1;
-    @java.lang.Override
-    public final boolean isInitialized() {
-      byte isInitialized = memoizedIsInitialized;
-      if (isInitialized == 1) return true;
-      if (isInitialized == 0) return false;
-
-      memoizedIsInitialized = 1;
-      return true;
-    }
-
-    @java.lang.Override
-    public void writeTo(com.google.protobuf.CodedOutputStream output)
-                        throws java.io.IOException {
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(id_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 1, id_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(namespace_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 2, namespace_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(fileName_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 3, fileName_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(fileUid_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 4, fileUid_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(path_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 5, path_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(mediaType_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 6, mediaType_);
-      }
-      if (purpose_ != 0) {
-        output.writeInt32(7, purpose_);
-      }
-      if (indexPolicy_ != 0) {
-        output.writeInt32(8, indexPolicy_);
-      }
-      if (retention_ != 0) {
-        output.writeInt32(9, retention_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(objectKey_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 10, objectKey_);
-      }
-      if (expectedSizeBytes_ != 0L) {
-        output.writeUInt64(11, expectedSizeBytes_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(expectedSha256_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 12, expectedSha256_);
-      }
-      com.google.protobuf.GeneratedMessage
-        .serializeStringMapTo(
-          output,
-          internalGetRequiredHeaders(),
-          RequiredHeadersDefaultEntryHolder.defaultEntry,
-          13);
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(createdByAgent_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 14, createdByAgent_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(createdBySessionId_)) {
-        com.google.protobuf.GeneratedMessage.writeString(output, 15, createdBySessionId_);
-      }
-      if (expiresAt_ != 0L) {
-        output.writeInt64(16, expiresAt_);
-      }
-      if (createdAt_ != 0L) {
-        output.writeInt64(17, createdAt_);
-      }
-      if (consumed_ != false) {
-        output.writeBool(18, consumed_);
-      }
-      getUnknownFields().writeTo(output);
-    }
-
-    @java.lang.Override
-    public int getSerializedSize() {
-      int size = memoizedSize;
-      if (size != -1) return size;
-
-      size = 0;
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(id_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(1, id_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(namespace_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(2, namespace_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(fileName_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(3, fileName_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(fileUid_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(4, fileUid_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(path_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(5, path_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(mediaType_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(6, mediaType_);
-      }
-      if (purpose_ != 0) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeInt32Size(7, purpose_);
-      }
-      if (indexPolicy_ != 0) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeInt32Size(8, indexPolicy_);
-      }
-      if (retention_ != 0) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeInt32Size(9, retention_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(objectKey_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(10, objectKey_);
-      }
-      if (expectedSizeBytes_ != 0L) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeUInt64Size(11, expectedSizeBytes_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(expectedSha256_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(12, expectedSha256_);
-      }
-      for (java.util.Map.Entry<java.lang.String, java.lang.String> entry
-           : internalGetRequiredHeaders().getMap().entrySet()) {
-        com.google.protobuf.MapEntry<java.lang.String, java.lang.String>
-        requiredHeaders__ = RequiredHeadersDefaultEntryHolder.defaultEntry.newBuilderForType()
-            .setKey(entry.getKey())
-            .setValue(entry.getValue())
-            .buildPartial();
-        size += com.google.protobuf.CodedOutputStream
-            .computeMessageSize(13, requiredHeaders__);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(createdByAgent_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(14, createdByAgent_);
-      }
-      if (!com.google.protobuf.GeneratedMessage.isStringEmpty(createdBySessionId_)) {
-        size += com.google.protobuf.GeneratedMessage.computeStringSize(15, createdBySessionId_);
-      }
-      if (expiresAt_ != 0L) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeInt64Size(16, expiresAt_);
-      }
-      if (createdAt_ != 0L) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeInt64Size(17, createdAt_);
-      }
-      if (consumed_ != false) {
-        size += com.google.protobuf.CodedOutputStream
-          .computeBoolSize(18, consumed_);
-      }
-      size += getUnknownFields().getSerializedSize();
-      memoizedSize = size;
-      return size;
-    }
-
-    @java.lang.Override
-    public boolean equals(final java.lang.Object obj) {
-      if (obj == this) {
-       return true;
-      }
-      if (!(obj instanceof talon.data.Data.FileUpload)) {
-        return super.equals(obj);
-      }
-      talon.data.Data.FileUpload other = (talon.data.Data.FileUpload) obj;
-
-      if (!getId()
-          .equals(other.getId())) return false;
-      if (!getNamespace()
-          .equals(other.getNamespace())) return false;
-      if (!getFileName()
-          .equals(other.getFileName())) return false;
-      if (!getFileUid()
-          .equals(other.getFileUid())) return false;
-      if (!getPath()
-          .equals(other.getPath())) return false;
-      if (!getMediaType()
-          .equals(other.getMediaType())) return false;
-      if (getPurpose()
-          != other.getPurpose()) return false;
-      if (getIndexPolicy()
-          != other.getIndexPolicy()) return false;
-      if (getRetention()
-          != other.getRetention()) return false;
-      if (!getObjectKey()
-          .equals(other.getObjectKey())) return false;
-      if (getExpectedSizeBytes()
-          != other.getExpectedSizeBytes()) return false;
-      if (!getExpectedSha256()
-          .equals(other.getExpectedSha256())) return false;
-      if (!internalGetRequiredHeaders().equals(
-          other.internalGetRequiredHeaders())) return false;
-      if (!getCreatedByAgent()
-          .equals(other.getCreatedByAgent())) return false;
-      if (!getCreatedBySessionId()
-          .equals(other.getCreatedBySessionId())) return false;
-      if (getExpiresAt()
-          != other.getExpiresAt()) return false;
-      if (getCreatedAt()
-          != other.getCreatedAt()) return false;
-      if (getConsumed()
-          != other.getConsumed()) return false;
-      if (!getUnknownFields().equals(other.getUnknownFields())) return false;
-      return true;
-    }
-
-    @java.lang.Override
-    public int hashCode() {
-      if (memoizedHashCode != 0) {
-        return memoizedHashCode;
-      }
-      int hash = 41;
-      hash = (19 * hash) + getDescriptor().hashCode();
-      hash = (37 * hash) + ID_FIELD_NUMBER;
-      hash = (53 * hash) + getId().hashCode();
-      hash = (37 * hash) + NAMESPACE_FIELD_NUMBER;
-      hash = (53 * hash) + getNamespace().hashCode();
-      hash = (37 * hash) + FILE_NAME_FIELD_NUMBER;
-      hash = (53 * hash) + getFileName().hashCode();
-      hash = (37 * hash) + FILE_UID_FIELD_NUMBER;
-      hash = (53 * hash) + getFileUid().hashCode();
-      hash = (37 * hash) + PATH_FIELD_NUMBER;
-      hash = (53 * hash) + getPath().hashCode();
-      hash = (37 * hash) + MEDIA_TYPE_FIELD_NUMBER;
-      hash = (53 * hash) + getMediaType().hashCode();
-      hash = (37 * hash) + PURPOSE_FIELD_NUMBER;
-      hash = (53 * hash) + getPurpose();
-      hash = (37 * hash) + INDEX_POLICY_FIELD_NUMBER;
-      hash = (53 * hash) + getIndexPolicy();
-      hash = (37 * hash) + RETENTION_FIELD_NUMBER;
-      hash = (53 * hash) + getRetention();
-      hash = (37 * hash) + OBJECT_KEY_FIELD_NUMBER;
-      hash = (53 * hash) + getObjectKey().hashCode();
-      hash = (37 * hash) + EXPECTED_SIZE_BYTES_FIELD_NUMBER;
-      hash = (53 * hash) + com.google.protobuf.Internal.hashLong(
-          getExpectedSizeBytes());
-      hash = (37 * hash) + EXPECTED_SHA256_FIELD_NUMBER;
-      hash = (53 * hash) + getExpectedSha256().hashCode();
-      if (!internalGetRequiredHeaders().getMap().isEmpty()) {
-        hash = (37 * hash) + REQUIRED_HEADERS_FIELD_NUMBER;
-        hash = (53 * hash) + internalGetRequiredHeaders().hashCode();
-      }
-      hash = (37 * hash) + CREATED_BY_AGENT_FIELD_NUMBER;
-      hash = (53 * hash) + getCreatedByAgent().hashCode();
-      hash = (37 * hash) + CREATED_BY_SESSION_ID_FIELD_NUMBER;
-      hash = (53 * hash) + getCreatedBySessionId().hashCode();
-      hash = (37 * hash) + EXPIRES_AT_FIELD_NUMBER;
-      hash = (53 * hash) + com.google.protobuf.Internal.hashLong(
-          getExpiresAt());
-      hash = (37 * hash) + CREATED_AT_FIELD_NUMBER;
-      hash = (53 * hash) + com.google.protobuf.Internal.hashLong(
-          getCreatedAt());
-      hash = (37 * hash) + CONSUMED_FIELD_NUMBER;
-      hash = (53 * hash) + com.google.protobuf.Internal.hashBoolean(
-          getConsumed());
-      hash = (29 * hash) + getUnknownFields().hashCode();
-      memoizedHashCode = hash;
-      return hash;
-    }
-
-    public static talon.data.Data.FileUpload parseFrom(
-        java.nio.ByteBuffer data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        java.nio.ByteBuffer data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        com.google.protobuf.ByteString data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        com.google.protobuf.ByteString data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static talon.data.Data.FileUpload parseFrom(byte[] data)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        byte[] data,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws com.google.protobuf.InvalidProtocolBufferException {
-      return PARSER.parseFrom(data, extensionRegistry);
-    }
-    public static talon.data.Data.FileUpload parseFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessage
-          .parseWithIOException(PARSER, input);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessage
-          .parseWithIOException(PARSER, input, extensionRegistry);
-    }
-
-    public static talon.data.Data.FileUpload parseDelimitedFrom(java.io.InputStream input)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessage
-          .parseDelimitedWithIOException(PARSER, input);
-    }
-
-    public static talon.data.Data.FileUpload parseDelimitedFrom(
-        java.io.InputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessage
-          .parseDelimitedWithIOException(PARSER, input, extensionRegistry);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        com.google.protobuf.CodedInputStream input)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessage
-          .parseWithIOException(PARSER, input);
-    }
-    public static talon.data.Data.FileUpload parseFrom(
-        com.google.protobuf.CodedInputStream input,
-        com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-        throws java.io.IOException {
-      return com.google.protobuf.GeneratedMessage
-          .parseWithIOException(PARSER, input, extensionRegistry);
-    }
-
-    @java.lang.Override
-    public Builder newBuilderForType() { return newBuilder(); }
-    public static Builder newBuilder() {
-      return DEFAULT_INSTANCE.toBuilder();
-    }
-    public static Builder newBuilder(talon.data.Data.FileUpload prototype) {
-      return DEFAULT_INSTANCE.toBuilder().mergeFrom(prototype);
-    }
-    @java.lang.Override
-    public Builder toBuilder() {
-      return this == DEFAULT_INSTANCE
-          ? new Builder() : new Builder().mergeFrom(this);
-    }
-
-    @java.lang.Override
-    protected Builder newBuilderForType(
-        com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-      Builder builder = new Builder(parent);
-      return builder;
-    }
-    /**
-     * <pre>
-     * Pending direct-to-object-storage upload for a namespace File.
-     *
-     * The target File is not committed to this object until CompleteFileUpload
-     * validates the uploaded object metadata and patches File.status.objectRef.
-     * </pre>
-     *
-     * Protobuf type {@code talon.data.FileUpload}
-     */
-    public static final class Builder extends
-        com.google.protobuf.GeneratedMessage.Builder<Builder> implements
-        // @@protoc_insertion_point(builder_implements:talon.data.FileUpload)
-        talon.data.Data.FileUploadOrBuilder {
-      public static final com.google.protobuf.Descriptors.Descriptor
-          getDescriptor() {
-        return talon.data.Data.internal_static_talon_data_FileUpload_descriptor;
-      }
-
-      @SuppressWarnings({"rawtypes"})
-      protected com.google.protobuf.MapFieldReflectionAccessor internalGetMapFieldReflection(
-          int number) {
-        switch (number) {
-          case 13:
-            return internalGetRequiredHeaders();
-          default:
-            throw new RuntimeException(
-                "Invalid map field number: " + number);
-        }
-      }
-      @SuppressWarnings({"rawtypes"})
-      protected com.google.protobuf.MapFieldReflectionAccessor internalGetMutableMapFieldReflection(
-          int number) {
-        switch (number) {
-          case 13:
-            return internalGetMutableRequiredHeaders();
-          default:
-            throw new RuntimeException(
-                "Invalid map field number: " + number);
-        }
-      }
-      @java.lang.Override
-      protected com.google.protobuf.GeneratedMessage.FieldAccessorTable
-          internalGetFieldAccessorTable() {
-        return talon.data.Data.internal_static_talon_data_FileUpload_fieldAccessorTable
-            .ensureFieldAccessorsInitialized(
-                talon.data.Data.FileUpload.class, talon.data.Data.FileUpload.Builder.class);
-      }
-
-      // Construct using talon.data.Data.FileUpload.newBuilder()
-      private Builder() {
-
-      }
-
-      private Builder(
-          com.google.protobuf.GeneratedMessage.BuilderParent parent) {
-        super(parent);
-
-      }
-      @java.lang.Override
-      public Builder clear() {
-        super.clear();
-        bitField0_ = 0;
-        id_ = "";
-        namespace_ = "";
-        fileName_ = "";
-        fileUid_ = "";
-        path_ = "";
-        mediaType_ = "";
-        purpose_ = 0;
-        indexPolicy_ = 0;
-        retention_ = 0;
-        objectKey_ = "";
-        expectedSizeBytes_ = 0L;
-        expectedSha256_ = "";
-        internalGetMutableRequiredHeaders().clear();
-        createdByAgent_ = "";
-        createdBySessionId_ = "";
-        expiresAt_ = 0L;
-        createdAt_ = 0L;
-        consumed_ = false;
-        return this;
-      }
-
-      @java.lang.Override
-      public com.google.protobuf.Descriptors.Descriptor
-          getDescriptorForType() {
-        return talon.data.Data.internal_static_talon_data_FileUpload_descriptor;
-      }
-
-      @java.lang.Override
-      public talon.data.Data.FileUpload getDefaultInstanceForType() {
-        return talon.data.Data.FileUpload.getDefaultInstance();
-      }
-
-      @java.lang.Override
-      public talon.data.Data.FileUpload build() {
-        talon.data.Data.FileUpload result = buildPartial();
-        if (!result.isInitialized()) {
-          throw newUninitializedMessageException(result);
-        }
-        return result;
-      }
-
-      @java.lang.Override
-      public talon.data.Data.FileUpload buildPartial() {
-        talon.data.Data.FileUpload result = new talon.data.Data.FileUpload(this);
-        if (bitField0_ != 0) { buildPartial0(result); }
-        onBuilt();
-        return result;
-      }
-
-      private void buildPartial0(talon.data.Data.FileUpload result) {
-        int from_bitField0_ = bitField0_;
-        if (((from_bitField0_ & 0x00000001) != 0)) {
-          result.id_ = id_;
-        }
-        if (((from_bitField0_ & 0x00000002) != 0)) {
-          result.namespace_ = namespace_;
-        }
-        if (((from_bitField0_ & 0x00000004) != 0)) {
-          result.fileName_ = fileName_;
-        }
-        if (((from_bitField0_ & 0x00000008) != 0)) {
-          result.fileUid_ = fileUid_;
-        }
-        if (((from_bitField0_ & 0x00000010) != 0)) {
-          result.path_ = path_;
-        }
-        if (((from_bitField0_ & 0x00000020) != 0)) {
-          result.mediaType_ = mediaType_;
-        }
-        if (((from_bitField0_ & 0x00000040) != 0)) {
-          result.purpose_ = purpose_;
-        }
-        if (((from_bitField0_ & 0x00000080) != 0)) {
-          result.indexPolicy_ = indexPolicy_;
-        }
-        if (((from_bitField0_ & 0x00000100) != 0)) {
-          result.retention_ = retention_;
-        }
-        if (((from_bitField0_ & 0x00000200) != 0)) {
-          result.objectKey_ = objectKey_;
-        }
-        if (((from_bitField0_ & 0x00000400) != 0)) {
-          result.expectedSizeBytes_ = expectedSizeBytes_;
-        }
-        if (((from_bitField0_ & 0x00000800) != 0)) {
-          result.expectedSha256_ = expectedSha256_;
-        }
-        if (((from_bitField0_ & 0x00001000) != 0)) {
-          result.requiredHeaders_ = internalGetRequiredHeaders();
-          result.requiredHeaders_.makeImmutable();
-        }
-        if (((from_bitField0_ & 0x00002000) != 0)) {
-          result.createdByAgent_ = createdByAgent_;
-        }
-        if (((from_bitField0_ & 0x00004000) != 0)) {
-          result.createdBySessionId_ = createdBySessionId_;
-        }
-        if (((from_bitField0_ & 0x00008000) != 0)) {
-          result.expiresAt_ = expiresAt_;
-        }
-        if (((from_bitField0_ & 0x00010000) != 0)) {
-          result.createdAt_ = createdAt_;
-        }
-        if (((from_bitField0_ & 0x00020000) != 0)) {
-          result.consumed_ = consumed_;
-        }
-      }
-
-      @java.lang.Override
-      public Builder mergeFrom(com.google.protobuf.Message other) {
-        if (other instanceof talon.data.Data.FileUpload) {
-          return mergeFrom((talon.data.Data.FileUpload)other);
-        } else {
-          super.mergeFrom(other);
-          return this;
-        }
-      }
-
-      public Builder mergeFrom(talon.data.Data.FileUpload other) {
-        if (other == talon.data.Data.FileUpload.getDefaultInstance()) return this;
-        if (!other.getId().isEmpty()) {
-          id_ = other.id_;
-          bitField0_ |= 0x00000001;
-          onChanged();
-        }
-        if (!other.getNamespace().isEmpty()) {
-          namespace_ = other.namespace_;
-          bitField0_ |= 0x00000002;
-          onChanged();
-        }
-        if (!other.getFileName().isEmpty()) {
-          fileName_ = other.fileName_;
-          bitField0_ |= 0x00000004;
-          onChanged();
-        }
-        if (!other.getFileUid().isEmpty()) {
-          fileUid_ = other.fileUid_;
-          bitField0_ |= 0x00000008;
-          onChanged();
-        }
-        if (!other.getPath().isEmpty()) {
-          path_ = other.path_;
-          bitField0_ |= 0x00000010;
-          onChanged();
-        }
-        if (!other.getMediaType().isEmpty()) {
-          mediaType_ = other.mediaType_;
-          bitField0_ |= 0x00000020;
-          onChanged();
-        }
-        if (other.getPurpose() != 0) {
-          setPurpose(other.getPurpose());
-        }
-        if (other.getIndexPolicy() != 0) {
-          setIndexPolicy(other.getIndexPolicy());
-        }
-        if (other.getRetention() != 0) {
-          setRetention(other.getRetention());
-        }
-        if (!other.getObjectKey().isEmpty()) {
-          objectKey_ = other.objectKey_;
-          bitField0_ |= 0x00000200;
-          onChanged();
-        }
-        if (other.getExpectedSizeBytes() != 0L) {
-          setExpectedSizeBytes(other.getExpectedSizeBytes());
-        }
-        if (!other.getExpectedSha256().isEmpty()) {
-          expectedSha256_ = other.expectedSha256_;
-          bitField0_ |= 0x00000800;
-          onChanged();
-        }
-        internalGetMutableRequiredHeaders().mergeFrom(
-            other.internalGetRequiredHeaders());
-        bitField0_ |= 0x00001000;
-        if (!other.getCreatedByAgent().isEmpty()) {
-          createdByAgent_ = other.createdByAgent_;
-          bitField0_ |= 0x00002000;
-          onChanged();
-        }
-        if (!other.getCreatedBySessionId().isEmpty()) {
-          createdBySessionId_ = other.createdBySessionId_;
-          bitField0_ |= 0x00004000;
-          onChanged();
-        }
-        if (other.getExpiresAt() != 0L) {
-          setExpiresAt(other.getExpiresAt());
-        }
-        if (other.getCreatedAt() != 0L) {
-          setCreatedAt(other.getCreatedAt());
-        }
-        if (other.getConsumed() != false) {
-          setConsumed(other.getConsumed());
-        }
-        this.mergeUnknownFields(other.getUnknownFields());
-        onChanged();
-        return this;
-      }
-
-      @java.lang.Override
-      public final boolean isInitialized() {
-        return true;
-      }
-
-      @java.lang.Override
-      public Builder mergeFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws java.io.IOException {
-        if (extensionRegistry == null) {
-          throw new java.lang.NullPointerException();
-        }
-        try {
-          boolean done = false;
-          while (!done) {
-            int tag = input.readTag();
-            switch (tag) {
-              case 0:
-                done = true;
-                break;
-              case 10: {
-                id_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000001;
-                break;
-              } // case 10
-              case 18: {
-                namespace_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000002;
-                break;
-              } // case 18
-              case 26: {
-                fileName_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000004;
-                break;
-              } // case 26
-              case 34: {
-                fileUid_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000008;
-                break;
-              } // case 34
-              case 42: {
-                path_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000010;
-                break;
-              } // case 42
-              case 50: {
-                mediaType_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000020;
-                break;
-              } // case 50
-              case 56: {
-                purpose_ = input.readInt32();
-                bitField0_ |= 0x00000040;
-                break;
-              } // case 56
-              case 64: {
-                indexPolicy_ = input.readInt32();
-                bitField0_ |= 0x00000080;
-                break;
-              } // case 64
-              case 72: {
-                retention_ = input.readInt32();
-                bitField0_ |= 0x00000100;
-                break;
-              } // case 72
-              case 82: {
-                objectKey_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000200;
-                break;
-              } // case 82
-              case 88: {
-                expectedSizeBytes_ = input.readUInt64();
-                bitField0_ |= 0x00000400;
-                break;
-              } // case 88
-              case 98: {
-                expectedSha256_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00000800;
-                break;
-              } // case 98
-              case 106: {
-                com.google.protobuf.MapEntry<java.lang.String, java.lang.String>
-                requiredHeaders__ = input.readMessage(
-                    RequiredHeadersDefaultEntryHolder.defaultEntry.getParserForType(), extensionRegistry);
-                internalGetMutableRequiredHeaders().getMutableMap().put(
-                    requiredHeaders__.getKey(), requiredHeaders__.getValue());
-                bitField0_ |= 0x00001000;
-                break;
-              } // case 106
-              case 114: {
-                createdByAgent_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00002000;
-                break;
-              } // case 114
-              case 122: {
-                createdBySessionId_ = input.readStringRequireUtf8();
-                bitField0_ |= 0x00004000;
-                break;
-              } // case 122
-              case 128: {
-                expiresAt_ = input.readInt64();
-                bitField0_ |= 0x00008000;
-                break;
-              } // case 128
-              case 136: {
-                createdAt_ = input.readInt64();
-                bitField0_ |= 0x00010000;
-                break;
-              } // case 136
-              case 144: {
-                consumed_ = input.readBool();
-                bitField0_ |= 0x00020000;
-                break;
-              } // case 144
-              default: {
-                if (!super.parseUnknownField(input, extensionRegistry, tag)) {
-                  done = true; // was an endgroup tag
-                }
-                break;
-              } // default:
-            } // switch (tag)
-          } // while (!done)
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          throw e.unwrapIOException();
-        } finally {
-          onChanged();
-        } // finally
-        return this;
-      }
-      private int bitField0_;
-
-      private java.lang.Object id_ = "";
-      /**
-       * <code>string id = 1;</code>
-       * @return The id.
-       */
-      public java.lang.String getId() {
-        java.lang.Object ref = id_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          id_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string id = 1;</code>
-       * @return The bytes for id.
-       */
-      public com.google.protobuf.ByteString
-          getIdBytes() {
-        java.lang.Object ref = id_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          id_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string id = 1;</code>
-       * @param value The id to set.
-       * @return This builder for chaining.
-       */
-      public Builder setId(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        id_ = value;
-        bitField0_ |= 0x00000001;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string id = 1;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearId() {
-        id_ = getDefaultInstance().getId();
-        bitField0_ = (bitField0_ & ~0x00000001);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string id = 1;</code>
-       * @param value The bytes for id to set.
-       * @return This builder for chaining.
-       */
-      public Builder setIdBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        id_ = value;
-        bitField0_ |= 0x00000001;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object namespace_ = "";
-      /**
-       * <code>string namespace = 2;</code>
-       * @return The namespace.
-       */
-      public java.lang.String getNamespace() {
-        java.lang.Object ref = namespace_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          namespace_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string namespace = 2;</code>
-       * @return The bytes for namespace.
-       */
-      public com.google.protobuf.ByteString
-          getNamespaceBytes() {
-        java.lang.Object ref = namespace_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          namespace_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string namespace = 2;</code>
-       * @param value The namespace to set.
-       * @return This builder for chaining.
-       */
-      public Builder setNamespace(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        namespace_ = value;
-        bitField0_ |= 0x00000002;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string namespace = 2;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearNamespace() {
-        namespace_ = getDefaultInstance().getNamespace();
-        bitField0_ = (bitField0_ & ~0x00000002);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string namespace = 2;</code>
-       * @param value The bytes for namespace to set.
-       * @return This builder for chaining.
-       */
-      public Builder setNamespaceBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        namespace_ = value;
-        bitField0_ |= 0x00000002;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object fileName_ = "";
-      /**
-       * <code>string file_name = 3;</code>
-       * @return The fileName.
-       */
-      public java.lang.String getFileName() {
-        java.lang.Object ref = fileName_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          fileName_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string file_name = 3;</code>
-       * @return The bytes for fileName.
-       */
-      public com.google.protobuf.ByteString
-          getFileNameBytes() {
-        java.lang.Object ref = fileName_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          fileName_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string file_name = 3;</code>
-       * @param value The fileName to set.
-       * @return This builder for chaining.
-       */
-      public Builder setFileName(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        fileName_ = value;
-        bitField0_ |= 0x00000004;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string file_name = 3;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearFileName() {
-        fileName_ = getDefaultInstance().getFileName();
-        bitField0_ = (bitField0_ & ~0x00000004);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string file_name = 3;</code>
-       * @param value The bytes for fileName to set.
-       * @return This builder for chaining.
-       */
-      public Builder setFileNameBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        fileName_ = value;
-        bitField0_ |= 0x00000004;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object fileUid_ = "";
-      /**
-       * <code>string file_uid = 4;</code>
-       * @return The fileUid.
-       */
-      public java.lang.String getFileUid() {
-        java.lang.Object ref = fileUid_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          fileUid_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string file_uid = 4;</code>
-       * @return The bytes for fileUid.
-       */
-      public com.google.protobuf.ByteString
-          getFileUidBytes() {
-        java.lang.Object ref = fileUid_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          fileUid_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string file_uid = 4;</code>
-       * @param value The fileUid to set.
-       * @return This builder for chaining.
-       */
-      public Builder setFileUid(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        fileUid_ = value;
-        bitField0_ |= 0x00000008;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string file_uid = 4;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearFileUid() {
-        fileUid_ = getDefaultInstance().getFileUid();
-        bitField0_ = (bitField0_ & ~0x00000008);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string file_uid = 4;</code>
-       * @param value The bytes for fileUid to set.
-       * @return This builder for chaining.
-       */
-      public Builder setFileUidBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        fileUid_ = value;
-        bitField0_ |= 0x00000008;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object path_ = "";
-      /**
-       * <code>string path = 5;</code>
-       * @return The path.
-       */
-      public java.lang.String getPath() {
-        java.lang.Object ref = path_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          path_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string path = 5;</code>
-       * @return The bytes for path.
-       */
-      public com.google.protobuf.ByteString
-          getPathBytes() {
-        java.lang.Object ref = path_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          path_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string path = 5;</code>
-       * @param value The path to set.
-       * @return This builder for chaining.
-       */
-      public Builder setPath(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        path_ = value;
-        bitField0_ |= 0x00000010;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string path = 5;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearPath() {
-        path_ = getDefaultInstance().getPath();
-        bitField0_ = (bitField0_ & ~0x00000010);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string path = 5;</code>
-       * @param value The bytes for path to set.
-       * @return This builder for chaining.
-       */
-      public Builder setPathBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        path_ = value;
-        bitField0_ |= 0x00000010;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object mediaType_ = "";
-      /**
-       * <code>string media_type = 6;</code>
-       * @return The mediaType.
-       */
-      public java.lang.String getMediaType() {
-        java.lang.Object ref = mediaType_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          mediaType_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string media_type = 6;</code>
-       * @return The bytes for mediaType.
-       */
-      public com.google.protobuf.ByteString
-          getMediaTypeBytes() {
-        java.lang.Object ref = mediaType_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          mediaType_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string media_type = 6;</code>
-       * @param value The mediaType to set.
-       * @return This builder for chaining.
-       */
-      public Builder setMediaType(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        mediaType_ = value;
-        bitField0_ |= 0x00000020;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string media_type = 6;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearMediaType() {
-        mediaType_ = getDefaultInstance().getMediaType();
-        bitField0_ = (bitField0_ & ~0x00000020);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string media_type = 6;</code>
-       * @param value The bytes for mediaType to set.
-       * @return This builder for chaining.
-       */
-      public Builder setMediaTypeBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        mediaType_ = value;
-        bitField0_ |= 0x00000020;
-        onChanged();
-        return this;
-      }
-
-      private int purpose_ ;
-      /**
-       * <code>int32 purpose = 7;</code>
-       * @return The purpose.
-       */
-      @java.lang.Override
-      public int getPurpose() {
-        return purpose_;
-      }
-      /**
-       * <code>int32 purpose = 7;</code>
-       * @param value The purpose to set.
-       * @return This builder for chaining.
-       */
-      public Builder setPurpose(int value) {
-
-        purpose_ = value;
-        bitField0_ |= 0x00000040;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>int32 purpose = 7;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearPurpose() {
-        bitField0_ = (bitField0_ & ~0x00000040);
-        purpose_ = 0;
-        onChanged();
-        return this;
-      }
-
-      private int indexPolicy_ ;
-      /**
-       * <code>int32 index_policy = 8;</code>
-       * @return The indexPolicy.
-       */
-      @java.lang.Override
-      public int getIndexPolicy() {
-        return indexPolicy_;
-      }
-      /**
-       * <code>int32 index_policy = 8;</code>
-       * @param value The indexPolicy to set.
-       * @return This builder for chaining.
-       */
-      public Builder setIndexPolicy(int value) {
-
-        indexPolicy_ = value;
-        bitField0_ |= 0x00000080;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>int32 index_policy = 8;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearIndexPolicy() {
-        bitField0_ = (bitField0_ & ~0x00000080);
-        indexPolicy_ = 0;
-        onChanged();
-        return this;
-      }
-
-      private int retention_ ;
-      /**
-       * <code>int32 retention = 9;</code>
-       * @return The retention.
-       */
-      @java.lang.Override
-      public int getRetention() {
-        return retention_;
-      }
-      /**
-       * <code>int32 retention = 9;</code>
-       * @param value The retention to set.
-       * @return This builder for chaining.
-       */
-      public Builder setRetention(int value) {
-
-        retention_ = value;
-        bitField0_ |= 0x00000100;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>int32 retention = 9;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearRetention() {
-        bitField0_ = (bitField0_ & ~0x00000100);
-        retention_ = 0;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object objectKey_ = "";
-      /**
-       * <code>string object_key = 10;</code>
-       * @return The objectKey.
-       */
-      public java.lang.String getObjectKey() {
-        java.lang.Object ref = objectKey_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          objectKey_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string object_key = 10;</code>
-       * @return The bytes for objectKey.
-       */
-      public com.google.protobuf.ByteString
-          getObjectKeyBytes() {
-        java.lang.Object ref = objectKey_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          objectKey_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string object_key = 10;</code>
-       * @param value The objectKey to set.
-       * @return This builder for chaining.
-       */
-      public Builder setObjectKey(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        objectKey_ = value;
-        bitField0_ |= 0x00000200;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string object_key = 10;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearObjectKey() {
-        objectKey_ = getDefaultInstance().getObjectKey();
-        bitField0_ = (bitField0_ & ~0x00000200);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string object_key = 10;</code>
-       * @param value The bytes for objectKey to set.
-       * @return This builder for chaining.
-       */
-      public Builder setObjectKeyBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        objectKey_ = value;
-        bitField0_ |= 0x00000200;
-        onChanged();
-        return this;
-      }
-
-      private long expectedSizeBytes_ ;
-      /**
-       * <code>uint64 expected_size_bytes = 11;</code>
-       * @return The expectedSizeBytes.
-       */
-      @java.lang.Override
-      public long getExpectedSizeBytes() {
-        return expectedSizeBytes_;
-      }
-      /**
-       * <code>uint64 expected_size_bytes = 11;</code>
-       * @param value The expectedSizeBytes to set.
-       * @return This builder for chaining.
-       */
-      public Builder setExpectedSizeBytes(long value) {
-
-        expectedSizeBytes_ = value;
-        bitField0_ |= 0x00000400;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>uint64 expected_size_bytes = 11;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearExpectedSizeBytes() {
-        bitField0_ = (bitField0_ & ~0x00000400);
-        expectedSizeBytes_ = 0L;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object expectedSha256_ = "";
-      /**
-       * <code>string expected_sha256 = 12;</code>
-       * @return The expectedSha256.
-       */
-      public java.lang.String getExpectedSha256() {
-        java.lang.Object ref = expectedSha256_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          expectedSha256_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string expected_sha256 = 12;</code>
-       * @return The bytes for expectedSha256.
-       */
-      public com.google.protobuf.ByteString
-          getExpectedSha256Bytes() {
-        java.lang.Object ref = expectedSha256_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          expectedSha256_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string expected_sha256 = 12;</code>
-       * @param value The expectedSha256 to set.
-       * @return This builder for chaining.
-       */
-      public Builder setExpectedSha256(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        expectedSha256_ = value;
-        bitField0_ |= 0x00000800;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string expected_sha256 = 12;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearExpectedSha256() {
-        expectedSha256_ = getDefaultInstance().getExpectedSha256();
-        bitField0_ = (bitField0_ & ~0x00000800);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string expected_sha256 = 12;</code>
-       * @param value The bytes for expectedSha256 to set.
-       * @return This builder for chaining.
-       */
-      public Builder setExpectedSha256Bytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        expectedSha256_ = value;
-        bitField0_ |= 0x00000800;
-        onChanged();
-        return this;
-      }
-
-      private com.google.protobuf.MapField<
-          java.lang.String, java.lang.String> requiredHeaders_;
-      private com.google.protobuf.MapField<java.lang.String, java.lang.String>
-          internalGetRequiredHeaders() {
-        if (requiredHeaders_ == null) {
-          return com.google.protobuf.MapField.emptyMapField(
-              RequiredHeadersDefaultEntryHolder.defaultEntry);
-        }
-        return requiredHeaders_;
-      }
-      private com.google.protobuf.MapField<java.lang.String, java.lang.String>
-          internalGetMutableRequiredHeaders() {
-        if (requiredHeaders_ == null) {
-          requiredHeaders_ = com.google.protobuf.MapField.newMapField(
-              RequiredHeadersDefaultEntryHolder.defaultEntry);
-        }
-        if (!requiredHeaders_.isMutable()) {
-          requiredHeaders_ = requiredHeaders_.copy();
-        }
-        bitField0_ |= 0x00001000;
-        onChanged();
-        return requiredHeaders_;
-      }
-      public int getRequiredHeadersCount() {
-        return internalGetRequiredHeaders().getMap().size();
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      @java.lang.Override
-      public boolean containsRequiredHeaders(
-          java.lang.String key) {
-        if (key == null) { throw new NullPointerException("map key"); }
-        return internalGetRequiredHeaders().getMap().containsKey(key);
-      }
-      /**
-       * Use {@link #getRequiredHeadersMap()} instead.
-       */
-      @java.lang.Override
-      @java.lang.Deprecated
-      public java.util.Map<java.lang.String, java.lang.String> getRequiredHeaders() {
-        return getRequiredHeadersMap();
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      @java.lang.Override
-      public java.util.Map<java.lang.String, java.lang.String> getRequiredHeadersMap() {
-        return internalGetRequiredHeaders().getMap();
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      @java.lang.Override
-      public /* nullable */
-java.lang.String getRequiredHeadersOrDefault(
-          java.lang.String key,
-          /* nullable */
-java.lang.String defaultValue) {
-        if (key == null) { throw new NullPointerException("map key"); }
-        java.util.Map<java.lang.String, java.lang.String> map =
-            internalGetRequiredHeaders().getMap();
-        return map.containsKey(key) ? map.get(key) : defaultValue;
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      @java.lang.Override
-      public java.lang.String getRequiredHeadersOrThrow(
-          java.lang.String key) {
-        if (key == null) { throw new NullPointerException("map key"); }
-        java.util.Map<java.lang.String, java.lang.String> map =
-            internalGetRequiredHeaders().getMap();
-        if (!map.containsKey(key)) {
-          throw new java.lang.IllegalArgumentException();
-        }
-        return map.get(key);
-      }
-      public Builder clearRequiredHeaders() {
-        bitField0_ = (bitField0_ & ~0x00001000);
-        internalGetMutableRequiredHeaders().getMutableMap()
-            .clear();
-        return this;
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      public Builder removeRequiredHeaders(
-          java.lang.String key) {
-        if (key == null) { throw new NullPointerException("map key"); }
-        internalGetMutableRequiredHeaders().getMutableMap()
-            .remove(key);
-        return this;
-      }
-      /**
-       * Use alternate mutation accessors instead.
-       */
-      @java.lang.Deprecated
-      public java.util.Map<java.lang.String, java.lang.String>
-          getMutableRequiredHeaders() {
-        bitField0_ |= 0x00001000;
-        return internalGetMutableRequiredHeaders().getMutableMap();
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      public Builder putRequiredHeaders(
-          java.lang.String key,
-          java.lang.String value) {
-        if (key == null) { throw new NullPointerException("map key"); }
-        if (value == null) { throw new NullPointerException("map value"); }
-        internalGetMutableRequiredHeaders().getMutableMap()
-            .put(key, value);
-        bitField0_ |= 0x00001000;
-        return this;
-      }
-      /**
-       * <code>map&lt;string, string&gt; required_headers = 13;</code>
-       */
-      public Builder putAllRequiredHeaders(
-          java.util.Map<java.lang.String, java.lang.String> values) {
-        internalGetMutableRequiredHeaders().getMutableMap()
-            .putAll(values);
-        bitField0_ |= 0x00001000;
-        return this;
-      }
-
-      private java.lang.Object createdByAgent_ = "";
-      /**
-       * <code>string created_by_agent = 14;</code>
-       * @return The createdByAgent.
-       */
-      public java.lang.String getCreatedByAgent() {
-        java.lang.Object ref = createdByAgent_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          createdByAgent_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string created_by_agent = 14;</code>
-       * @return The bytes for createdByAgent.
-       */
-      public com.google.protobuf.ByteString
-          getCreatedByAgentBytes() {
-        java.lang.Object ref = createdByAgent_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          createdByAgent_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string created_by_agent = 14;</code>
-       * @param value The createdByAgent to set.
-       * @return This builder for chaining.
-       */
-      public Builder setCreatedByAgent(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        createdByAgent_ = value;
-        bitField0_ |= 0x00002000;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string created_by_agent = 14;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearCreatedByAgent() {
-        createdByAgent_ = getDefaultInstance().getCreatedByAgent();
-        bitField0_ = (bitField0_ & ~0x00002000);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string created_by_agent = 14;</code>
-       * @param value The bytes for createdByAgent to set.
-       * @return This builder for chaining.
-       */
-      public Builder setCreatedByAgentBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        createdByAgent_ = value;
-        bitField0_ |= 0x00002000;
-        onChanged();
-        return this;
-      }
-
-      private java.lang.Object createdBySessionId_ = "";
-      /**
-       * <code>string created_by_session_id = 15;</code>
-       * @return The createdBySessionId.
-       */
-      public java.lang.String getCreatedBySessionId() {
-        java.lang.Object ref = createdBySessionId_;
-        if (!(ref instanceof java.lang.String)) {
-          com.google.protobuf.ByteString bs =
-              (com.google.protobuf.ByteString) ref;
-          java.lang.String s = bs.toStringUtf8();
-          createdBySessionId_ = s;
-          return s;
-        } else {
-          return (java.lang.String) ref;
-        }
-      }
-      /**
-       * <code>string created_by_session_id = 15;</code>
-       * @return The bytes for createdBySessionId.
-       */
-      public com.google.protobuf.ByteString
-          getCreatedBySessionIdBytes() {
-        java.lang.Object ref = createdBySessionId_;
-        if (ref instanceof String) {
-          com.google.protobuf.ByteString b =
-              com.google.protobuf.ByteString.copyFromUtf8(
-                  (java.lang.String) ref);
-          createdBySessionId_ = b;
-          return b;
-        } else {
-          return (com.google.protobuf.ByteString) ref;
-        }
-      }
-      /**
-       * <code>string created_by_session_id = 15;</code>
-       * @param value The createdBySessionId to set.
-       * @return This builder for chaining.
-       */
-      public Builder setCreatedBySessionId(
-          java.lang.String value) {
-        if (value == null) { throw new NullPointerException(); }
-        createdBySessionId_ = value;
-        bitField0_ |= 0x00004000;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string created_by_session_id = 15;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearCreatedBySessionId() {
-        createdBySessionId_ = getDefaultInstance().getCreatedBySessionId();
-        bitField0_ = (bitField0_ & ~0x00004000);
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>string created_by_session_id = 15;</code>
-       * @param value The bytes for createdBySessionId to set.
-       * @return This builder for chaining.
-       */
-      public Builder setCreatedBySessionIdBytes(
-          com.google.protobuf.ByteString value) {
-        if (value == null) { throw new NullPointerException(); }
-        checkByteStringIsUtf8(value);
-        createdBySessionId_ = value;
-        bitField0_ |= 0x00004000;
-        onChanged();
-        return this;
-      }
-
-      private long expiresAt_ ;
-      /**
-       * <code>int64 expires_at = 16;</code>
-       * @return The expiresAt.
-       */
-      @java.lang.Override
-      public long getExpiresAt() {
-        return expiresAt_;
-      }
-      /**
-       * <code>int64 expires_at = 16;</code>
-       * @param value The expiresAt to set.
-       * @return This builder for chaining.
-       */
-      public Builder setExpiresAt(long value) {
-
-        expiresAt_ = value;
-        bitField0_ |= 0x00008000;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>int64 expires_at = 16;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearExpiresAt() {
-        bitField0_ = (bitField0_ & ~0x00008000);
-        expiresAt_ = 0L;
-        onChanged();
-        return this;
-      }
-
-      private long createdAt_ ;
-      /**
-       * <code>int64 created_at = 17;</code>
-       * @return The createdAt.
-       */
-      @java.lang.Override
-      public long getCreatedAt() {
-        return createdAt_;
-      }
-      /**
-       * <code>int64 created_at = 17;</code>
-       * @param value The createdAt to set.
-       * @return This builder for chaining.
-       */
-      public Builder setCreatedAt(long value) {
-
-        createdAt_ = value;
-        bitField0_ |= 0x00010000;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>int64 created_at = 17;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearCreatedAt() {
-        bitField0_ = (bitField0_ & ~0x00010000);
-        createdAt_ = 0L;
-        onChanged();
-        return this;
-      }
-
-      private boolean consumed_ ;
-      /**
-       * <code>bool consumed = 18;</code>
-       * @return The consumed.
-       */
-      @java.lang.Override
-      public boolean getConsumed() {
-        return consumed_;
-      }
-      /**
-       * <code>bool consumed = 18;</code>
-       * @param value The consumed to set.
-       * @return This builder for chaining.
-       */
-      public Builder setConsumed(boolean value) {
-
-        consumed_ = value;
-        bitField0_ |= 0x00020000;
-        onChanged();
-        return this;
-      }
-      /**
-       * <code>bool consumed = 18;</code>
-       * @return This builder for chaining.
-       */
-      public Builder clearConsumed() {
-        bitField0_ = (bitField0_ & ~0x00020000);
-        consumed_ = false;
-        onChanged();
-        return this;
-      }
-
-      // @@protoc_insertion_point(builder_scope:talon.data.FileUpload)
-    }
-
-    // @@protoc_insertion_point(class_scope:talon.data.FileUpload)
-    private static final talon.data.Data.FileUpload DEFAULT_INSTANCE;
-    static {
-      DEFAULT_INSTANCE = new talon.data.Data.FileUpload();
-    }
-
-    public static talon.data.Data.FileUpload getDefaultInstance() {
-      return DEFAULT_INSTANCE;
-    }
-
-    private static final com.google.protobuf.Parser<FileUpload>
-        PARSER = new com.google.protobuf.AbstractParser<FileUpload>() {
-      @java.lang.Override
-      public FileUpload parsePartialFrom(
-          com.google.protobuf.CodedInputStream input,
-          com.google.protobuf.ExtensionRegistryLite extensionRegistry)
-          throws com.google.protobuf.InvalidProtocolBufferException {
-        Builder builder = newBuilder();
-        try {
-          builder.mergeFrom(input, extensionRegistry);
-        } catch (com.google.protobuf.InvalidProtocolBufferException e) {
-          throw e.setUnfinishedMessage(builder.buildPartial());
-        } catch (com.google.protobuf.UninitializedMessageException e) {
-          throw e.asInvalidProtocolBufferException().setUnfinishedMessage(builder.buildPartial());
-        } catch (java.io.IOException e) {
-          throw new com.google.protobuf.InvalidProtocolBufferException(e)
-              .setUnfinishedMessage(builder.buildPartial());
-        }
-        return builder.buildPartial();
-      }
-    };
-
-    public static com.google.protobuf.Parser<FileUpload> parser() {
-      return PARSER;
-    }
-
-    @java.lang.Override
-    public com.google.protobuf.Parser<FileUpload> getParserForType() {
-      return PARSER;
-    }
-
-    @java.lang.Override
-    public talon.data.Data.FileUpload getDefaultInstanceForType() {
-      return DEFAULT_INSTANCE;
-    }
-
-  }
-
   public interface HandleGrantOrBuilder extends
       // @@protoc_insertion_point(interface_extends:talon.data.HandleGrant)
       com.google.protobuf.MessageOrBuilder {
@@ -26542,16 +23860,6 @@ java.lang.String defaultValue) {
     com.google.protobuf.GeneratedMessage.FieldAccessorTable
       internal_static_talon_data_Artifact_MetadataEntry_fieldAccessorTable;
   private static final com.google.protobuf.Descriptors.Descriptor
-    internal_static_talon_data_FileUpload_descriptor;
-  private static final
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_talon_data_FileUpload_fieldAccessorTable;
-  private static final com.google.protobuf.Descriptors.Descriptor
-    internal_static_talon_data_FileUpload_RequiredHeadersEntry_descriptor;
-  private static final
-    com.google.protobuf.GeneratedMessage.FieldAccessorTable
-      internal_static_talon_data_FileUpload_RequiredHeadersEntry_fieldAccessorTable;
-  private static final com.google.protobuf.Descriptors.Descriptor
     internal_static_talon_data_HandleGrant_descriptor;
   private static final
     com.google.protobuf.GeneratedMessage.FieldAccessorTable
@@ -26661,104 +23969,91 @@ java.lang.String defaultValue) {
       "talon.data.Artifact.MetadataEntry\032-\n\013Lab" +
       "elsEntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001" +
       "\032/\n\rMetadataEntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030" +
-      "\002 \001(\t:\0028\001\"\350\003\n\nFileUpload\022\n\n\002id\030\001 \001(\t\022\021\n\t" +
-      "namespace\030\002 \001(\t\022\021\n\tfile_name\030\003 \001(\t\022\020\n\010fi" +
-      "le_uid\030\004 \001(\t\022\014\n\004path\030\005 \001(\t\022\022\n\nmedia_type" +
-      "\030\006 \001(\t\022\017\n\007purpose\030\007 \001(\005\022\024\n\014index_policy\030" +
-      "\010 \001(\005\022\021\n\tretention\030\t \001(\005\022\022\n\nobject_key\030\n" +
-      " \001(\t\022\033\n\023expected_size_bytes\030\013 \001(\004\022\027\n\017exp" +
-      "ected_sha256\030\014 \001(\t\022E\n\020required_headers\030\r" +
-      " \003(\0132+.talon.data.FileUpload.RequiredHea" +
-      "dersEntry\022\030\n\020created_by_agent\030\016 \001(\t\022\035\n\025c" +
-      "reated_by_session_id\030\017 \001(\t\022\022\n\nexpires_at" +
-      "\030\020 \001(\003\022\022\n\ncreated_at\030\021 \001(\003\022\020\n\010consumed\030\022" +
-      " \001(\010\0326\n\024RequiredHeadersEntry\022\013\n\003key\030\001 \001(" +
-      "\t\022\r\n\005value\030\002 \001(\t:\0028\001\"\341\001\n\013HandleGrant\022\n\n\002" +
-      "id\030\001 \001(\t\022\021\n\tnamespace\030\002 \001(\t\022\014\n\004kind\030\003 \001(" +
-      "\t\022\021\n\ttarget_id\030\004 \001(\t\022\r\n\005agent\030\005 \001(\t\022\022\n\ns" +
-      "ession_id\030\006 \001(\t\022\022\n\noperations\030\007 \003(\t\022\026\n\016a" +
-      "udience_agent\030\010 \001(\t\022\033\n\023audience_session_" +
-      "id\030\t \001(\t\022\022\n\nexpires_at\030\n \001(\003\022\022\n\ncreated_" +
-      "at\030\013 \001(\003\"\275\001\n\tPrincipal\022\023\n\013external_id\030\001 " +
-      "\001(\t\022\017\n\007address\030\002 \001(\t\022\024\n\014display_name\030\003 \001" +
-      "(\t\022\014\n\004kind\030\004 \001(\t\0225\n\010metadata\030\005 \003(\0132#.tal" +
-      "on.data.Principal.MetadataEntry\032/\n\rMetad" +
-      "ataEntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001" +
-      "\"\307\001\n\022SessionMessagePart\022\n\n\002id\030\001 \001(\t\0225\n\tp" +
-      "art_type\030\002 \001(\0162\".talon.data.SessionMessa" +
-      "gePartType\022\017\n\007content\030\003 \001(\t\022\014\n\004name\030\004 \001(" +
-      "\t\022\024\n\014payload_json\030\005 \001(\t\022\022\n\ncreated_at\030\006 " +
-      "\001(\003\022%\n\006object\030\007 \001(\0132\025.talon.data.ObjectR" +
-      "ef\"\363\001\n\016SessionMessage\022\n\n\002id\030\001 \001(\t\022%\n\004rol" +
-      "e\030\002 \001(\0162\027.talon.data.MessageRole\022\022\n\ncrea" +
-      "ted_at\030\004 \001(\003\0226\n\006labels\030\005 \003(\0132&.talon.dat" +
-      "a.SessionMessage.LabelsEntry\022-\n\005parts\030\006 " +
-      "\003(\0132\036.talon.data.SessionMessagePart\032-\n\013L" +
-      "abelsEntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\002" +
-      "8\001J\004\010\003\020\004\"\257\002\n\007Session\022\n\n\002id\030\001 \001(\t\022\r\n\005agen" +
-      "t\030\002 \001(\t\022\n\n\002ns\030\003 \001(\t\022\016\n\006status\030\004 \001(\t\022\022\n\nc" +
-      "reated_at\030\005 \001(\003\022\023\n\013last_active\030\006 \001(\003\0223\n\010" +
-      "metadata\030\007 \003(\0132!.talon.data.Session.Meta" +
-      "dataEntry\022/\n\006labels\030\010 \003(\0132\037.talon.data.S" +
-      "ession.LabelsEntry\032/\n\rMetadataEntry\022\013\n\003k" +
-      "ey\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001\032-\n\013LabelsEnt" +
-      "ry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001\"\233\002\n\016C" +
-      "hannelMessage\022\n\n\002id\030\001 \001(\t\022\n\n\002ns\030\002 \001(\t\022\017\n" +
-      "\007channel\030\003 \001(\t\022\023\n\013author_kind\030\004 \001(\t\022\016\n\006a" +
-      "uthor\030\005 \001(\t\022\017\n\007content\030\006 \001(\t\022\022\n\ncreated_" +
-      "at\030\007 \001(\003\022\024\n\014source_agent\030\010 \001(\t\022\031\n\021source" +
-      "_session_id\030\t \001(\t\0226\n\006labels\030\n \003(\0132&.talo" +
-      "n.data.ChannelMessage.LabelsEntry\032-\n\013Lab" +
-      "elsEntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001" +
-      "\"_\n\tKnowledge\022\014\n\004path\030\001 \001(\t\022\017\n\007content\030\002" +
-      " \001(\t\022\022\n\nupdated_at\030\003 \001(\003\022\021\n\tnamespace\030\004 " +
-      "\001(\t\022\014\n\004name\030\005 \001(\t\"k\n\025KnowledgeSearchResu" +
-      "lt\022\014\n\004path\030\001 \001(\t\022\017\n\007snippet\030\002 \001(\t\022\r\n\005sco" +
-      "re\030\003 \001(\002\022\021\n\ttimestamp\030\004 \001(\003\022\021\n\tnamespace" +
-      "\030\005 \001(\t\"\313\003\n\013WorkflowRun\022\n\n\002id\030\001 \001(\t\022\020\n\010wo" +
-      "rkflow\030\002 \001(\t\022\n\n\002ns\030\003 \001(\t\022\016\n\006status\030\004 \001(\t" +
-      "\022\022\n\ninput_json\030\005 \001(\t\022\022\n\nstate_json\030\006 \001(\t" +
-      "\022\023\n\013output_json\030\007 \001(\t\022\022\n\ncreated_at\030\010 \001(" +
-      "\003\022\022\n\nupdated_at\030\t \001(\003\0223\n\006labels\030\n \003(\0132#." +
-      "talon.data.WorkflowRun.LabelsEntry\022\035\n\020cl" +
-      "aim_expires_at\030\013 \001(\003H\000\210\001\001\022\r\n\005error\030\014 \001(\t" +
-      "\022\021\n\tspec_json\030\r \001(\t\022\031\n\021workflow_revision" +
-      "\030\016 \001(\004\022\023\n\013claim_owner\030\017 \001(\t\022\025\n\rclaim_att" +
-      "empt\030\020 \001(\r\022\034\n\024last_dispatch_reason\030\021 \001(\t" +
-      "\032-\n\013LabelsEntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 " +
-      "\001(\t:\0028\001B\023\n\021_claim_expires_at\"\263\003\n\017Workflo" +
-      "wStepRun\022\n\n\002id\030\001 \001(\t\022\017\n\007step_id\030\002 \001(\t\022\017\n" +
-      "\007attempt\030\003 \001(\r\022\016\n\006status\030\004 \001(\t\022\022\n\ninput_" +
-      "json\030\005 \001(\t\022\023\n\013output_json\030\006 \001(\t\022\r\n\005error" +
-      "\030\007 \001(\t\022\030\n\020child_session_id\030\010 \001(\t\022\035\n\025chil" +
-      "d_workflow_run_id\030\t \001(\t\022\023\n\013resume_json\030\n" +
-      " \001(\t\022\024\n\014suspend_json\030\013 \001(\t\022\022\n\ncreated_at" +
-      "\030\014 \001(\003\022\022\n\nupdated_at\030\r \001(\003\022\032\n\rnext_retry" +
-      "_at\030\016 \001(\003H\000\210\001\001\022\027\n\ntimeout_at\030\017 \001(\003H\001\210\001\001\022" +
-      "\032\n\022wait_wakeup_handle\030\020 \001(\t\022\032\n\rwait_unti" +
-      "l_at\030\021 \001(\003H\002\210\001\001B\020\n\016_next_retry_atB\r\n\013_ti" +
-      "meout_atB\020\n\016_wait_until_at\"\245\001\n\020WorkflowR" +
-      "unEvent\022\n\n\002id\030\001 \001(\t\022\n\n\002ns\030\002 \001(\t\022\020\n\010workf" +
-      "low\030\003 \001(\t\022\016\n\006run_id\030\004 \001(\t\022\014\n\004type\030\005 \001(\t\022" +
-      "\017\n\007step_id\030\006 \001(\t\022\017\n\007message\030\007 \001(\t\022\024\n\014pay" +
-      "load_json\030\010 \001(\t\022\021\n\ttimestamp\030\t \001(\003*W\n\013Me" +
-      "ssageRole\022\024\n\020ROLE_UNSPECIFIED\020\000\022\r\n\tROLE_" +
-      "USER\020\001\022\022\n\016ROLE_ASSISTANT\020\002\022\017\n\013ROLE_SYSTE" +
-      "M\020\003*\244\004\n\026SessionMessagePartType\022)\n%SESSIO" +
-      "N_MESSAGE_PART_TYPE_UNSPECIFIED\020\000\022\"\n\036SES" +
-      "SION_MESSAGE_PART_TYPE_TEXT\020\001\022\'\n#SESSION" +
-      "_MESSAGE_PART_TYPE_REASONING\020\002\022\'\n#SESSIO" +
-      "N_MESSAGE_PART_TYPE_TOOL_CALL\020\003\022)\n%SESSI" +
-      "ON_MESSAGE_PART_TYPE_TOOL_RESULT\020\004\022#\n\037SE" +
-      "SSION_MESSAGE_PART_TYPE_USAGE\020\005\022#\n\037SESSI" +
-      "ON_MESSAGE_PART_TYPE_ERROR\020\006\022#\n\037SESSION_" +
-      "MESSAGE_PART_TYPE_IMAGE\020\007\022#\n\037SESSION_MES" +
-      "SAGE_PART_TYPE_AUDIO\020\010\022#\n\037SESSION_MESSAG" +
-      "E_PART_TYPE_VIDEO\020\t\022\"\n\036SESSION_MESSAGE_P" +
-      "ART_TYPE_FILE\020\n\0220\n,SESSION_MESSAGE_PART_" +
-      "TYPE_REQUEST_PERMISSION\020\013\022/\n+SESSION_MES" +
-      "SAGE_PART_TYPE_PERMISSION_RESULT\020\014b\006prot" +
-      "o3"
+      "\002 \001(\t:\0028\001\"\341\001\n\013HandleGrant\022\n\n\002id\030\001 \001(\t\022\021\n" +
+      "\tnamespace\030\002 \001(\t\022\014\n\004kind\030\003 \001(\t\022\021\n\ttarget" +
+      "_id\030\004 \001(\t\022\r\n\005agent\030\005 \001(\t\022\022\n\nsession_id\030\006" +
+      " \001(\t\022\022\n\noperations\030\007 \003(\t\022\026\n\016audience_age" +
+      "nt\030\010 \001(\t\022\033\n\023audience_session_id\030\t \001(\t\022\022\n" +
+      "\nexpires_at\030\n \001(\003\022\022\n\ncreated_at\030\013 \001(\003\"\275\001" +
+      "\n\tPrincipal\022\023\n\013external_id\030\001 \001(\t\022\017\n\007addr" +
+      "ess\030\002 \001(\t\022\024\n\014display_name\030\003 \001(\t\022\014\n\004kind\030" +
+      "\004 \001(\t\0225\n\010metadata\030\005 \003(\0132#.talon.data.Pri" +
+      "ncipal.MetadataEntry\032/\n\rMetadataEntry\022\013\n" +
+      "\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001\"\307\001\n\022Sessio" +
+      "nMessagePart\022\n\n\002id\030\001 \001(\t\0225\n\tpart_type\030\002 " +
+      "\001(\0162\".talon.data.SessionMessagePartType\022" +
+      "\017\n\007content\030\003 \001(\t\022\014\n\004name\030\004 \001(\t\022\024\n\014payloa" +
+      "d_json\030\005 \001(\t\022\022\n\ncreated_at\030\006 \001(\003\022%\n\006obje" +
+      "ct\030\007 \001(\0132\025.talon.data.ObjectRef\"\363\001\n\016Sess" +
+      "ionMessage\022\n\n\002id\030\001 \001(\t\022%\n\004role\030\002 \001(\0162\027.t" +
+      "alon.data.MessageRole\022\022\n\ncreated_at\030\004 \001(" +
+      "\003\0226\n\006labels\030\005 \003(\0132&.talon.data.SessionMe" +
+      "ssage.LabelsEntry\022-\n\005parts\030\006 \003(\0132\036.talon" +
+      ".data.SessionMessagePart\032-\n\013LabelsEntry\022" +
+      "\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001J\004\010\003\020\004\"\257\002" +
+      "\n\007Session\022\n\n\002id\030\001 \001(\t\022\r\n\005agent\030\002 \001(\t\022\n\n\002" +
+      "ns\030\003 \001(\t\022\016\n\006status\030\004 \001(\t\022\022\n\ncreated_at\030\005" +
+      " \001(\003\022\023\n\013last_active\030\006 \001(\003\0223\n\010metadata\030\007 " +
+      "\003(\0132!.talon.data.Session.MetadataEntry\022/" +
+      "\n\006labels\030\010 \003(\0132\037.talon.data.Session.Labe" +
+      "lsEntry\032/\n\rMetadataEntry\022\013\n\003key\030\001 \001(\t\022\r\n" +
+      "\005value\030\002 \001(\t:\0028\001\032-\n\013LabelsEntry\022\013\n\003key\030\001" +
+      " \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001\"\233\002\n\016ChannelMessa" +
+      "ge\022\n\n\002id\030\001 \001(\t\022\n\n\002ns\030\002 \001(\t\022\017\n\007channel\030\003 " +
+      "\001(\t\022\023\n\013author_kind\030\004 \001(\t\022\016\n\006author\030\005 \001(\t" +
+      "\022\017\n\007content\030\006 \001(\t\022\022\n\ncreated_at\030\007 \001(\003\022\024\n" +
+      "\014source_agent\030\010 \001(\t\022\031\n\021source_session_id" +
+      "\030\t \001(\t\0226\n\006labels\030\n \003(\0132&.talon.data.Chan" +
+      "nelMessage.LabelsEntry\032-\n\013LabelsEntry\022\013\n" +
+      "\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001\"_\n\tKnowled" +
+      "ge\022\014\n\004path\030\001 \001(\t\022\017\n\007content\030\002 \001(\t\022\022\n\nupd" +
+      "ated_at\030\003 \001(\003\022\021\n\tnamespace\030\004 \001(\t\022\014\n\004name" +
+      "\030\005 \001(\t\"k\n\025KnowledgeSearchResult\022\014\n\004path\030" +
+      "\001 \001(\t\022\017\n\007snippet\030\002 \001(\t\022\r\n\005score\030\003 \001(\002\022\021\n" +
+      "\ttimestamp\030\004 \001(\003\022\021\n\tnamespace\030\005 \001(\t\"\313\003\n\013" +
+      "WorkflowRun\022\n\n\002id\030\001 \001(\t\022\020\n\010workflow\030\002 \001(" +
+      "\t\022\n\n\002ns\030\003 \001(\t\022\016\n\006status\030\004 \001(\t\022\022\n\ninput_j" +
+      "son\030\005 \001(\t\022\022\n\nstate_json\030\006 \001(\t\022\023\n\013output_" +
+      "json\030\007 \001(\t\022\022\n\ncreated_at\030\010 \001(\003\022\022\n\nupdate" +
+      "d_at\030\t \001(\003\0223\n\006labels\030\n \003(\0132#.talon.data." +
+      "WorkflowRun.LabelsEntry\022\035\n\020claim_expires" +
+      "_at\030\013 \001(\003H\000\210\001\001\022\r\n\005error\030\014 \001(\t\022\021\n\tspec_js" +
+      "on\030\r \001(\t\022\031\n\021workflow_revision\030\016 \001(\004\022\023\n\013c" +
+      "laim_owner\030\017 \001(\t\022\025\n\rclaim_attempt\030\020 \001(\r\022" +
+      "\034\n\024last_dispatch_reason\030\021 \001(\t\032-\n\013LabelsE" +
+      "ntry\022\013\n\003key\030\001 \001(\t\022\r\n\005value\030\002 \001(\t:\0028\001B\023\n\021" +
+      "_claim_expires_at\"\263\003\n\017WorkflowStepRun\022\n\n" +
+      "\002id\030\001 \001(\t\022\017\n\007step_id\030\002 \001(\t\022\017\n\007attempt\030\003 " +
+      "\001(\r\022\016\n\006status\030\004 \001(\t\022\022\n\ninput_json\030\005 \001(\t\022" +
+      "\023\n\013output_json\030\006 \001(\t\022\r\n\005error\030\007 \001(\t\022\030\n\020c" +
+      "hild_session_id\030\010 \001(\t\022\035\n\025child_workflow_" +
+      "run_id\030\t \001(\t\022\023\n\013resume_json\030\n \001(\t\022\024\n\014sus" +
+      "pend_json\030\013 \001(\t\022\022\n\ncreated_at\030\014 \001(\003\022\022\n\nu" +
+      "pdated_at\030\r \001(\003\022\032\n\rnext_retry_at\030\016 \001(\003H\000" +
+      "\210\001\001\022\027\n\ntimeout_at\030\017 \001(\003H\001\210\001\001\022\032\n\022wait_wak" +
+      "eup_handle\030\020 \001(\t\022\032\n\rwait_until_at\030\021 \001(\003H" +
+      "\002\210\001\001B\020\n\016_next_retry_atB\r\n\013_timeout_atB\020\n" +
+      "\016_wait_until_at\"\245\001\n\020WorkflowRunEvent\022\n\n\002" +
+      "id\030\001 \001(\t\022\n\n\002ns\030\002 \001(\t\022\020\n\010workflow\030\003 \001(\t\022\016" +
+      "\n\006run_id\030\004 \001(\t\022\014\n\004type\030\005 \001(\t\022\017\n\007step_id\030" +
+      "\006 \001(\t\022\017\n\007message\030\007 \001(\t\022\024\n\014payload_json\030\010" +
+      " \001(\t\022\021\n\ttimestamp\030\t \001(\003*W\n\013MessageRole\022\024" +
+      "\n\020ROLE_UNSPECIFIED\020\000\022\r\n\tROLE_USER\020\001\022\022\n\016R" +
+      "OLE_ASSISTANT\020\002\022\017\n\013ROLE_SYSTEM\020\003*\244\004\n\026Ses" +
+      "sionMessagePartType\022)\n%SESSION_MESSAGE_P" +
+      "ART_TYPE_UNSPECIFIED\020\000\022\"\n\036SESSION_MESSAG" +
+      "E_PART_TYPE_TEXT\020\001\022\'\n#SESSION_MESSAGE_PA" +
+      "RT_TYPE_REASONING\020\002\022\'\n#SESSION_MESSAGE_P" +
+      "ART_TYPE_TOOL_CALL\020\003\022)\n%SESSION_MESSAGE_" +
+      "PART_TYPE_TOOL_RESULT\020\004\022#\n\037SESSION_MESSA" +
+      "GE_PART_TYPE_USAGE\020\005\022#\n\037SESSION_MESSAGE_" +
+      "PART_TYPE_ERROR\020\006\022#\n\037SESSION_MESSAGE_PAR" +
+      "T_TYPE_IMAGE\020\007\022#\n\037SESSION_MESSAGE_PART_T" +
+      "YPE_AUDIO\020\010\022#\n\037SESSION_MESSAGE_PART_TYPE" +
+      "_VIDEO\020\t\022\"\n\036SESSION_MESSAGE_PART_TYPE_FI" +
+      "LE\020\n\0220\n,SESSION_MESSAGE_PART_TYPE_REQUES" +
+      "T_PERMISSION\020\013\022/\n+SESSION_MESSAGE_PART_T" +
+      "YPE_PERMISSION_RESULT\020\014b\006proto3"
     };
     descriptor = com.google.protobuf.Descriptors.FileDescriptor
       .internalBuildGeneratedFileFrom(descriptorData,
@@ -26794,26 +24089,14 @@ java.lang.String defaultValue) {
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_Artifact_MetadataEntry_descriptor,
         new java.lang.String[] { "Key", "Value", });
-    internal_static_talon_data_FileUpload_descriptor =
-      getDescriptor().getMessageType(2);
-    internal_static_talon_data_FileUpload_fieldAccessorTable = new
-      com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-        internal_static_talon_data_FileUpload_descriptor,
-        new java.lang.String[] { "Id", "Namespace", "FileName", "FileUid", "Path", "MediaType", "Purpose", "IndexPolicy", "Retention", "ObjectKey", "ExpectedSizeBytes", "ExpectedSha256", "RequiredHeaders", "CreatedByAgent", "CreatedBySessionId", "ExpiresAt", "CreatedAt", "Consumed", });
-    internal_static_talon_data_FileUpload_RequiredHeadersEntry_descriptor =
-      internal_static_talon_data_FileUpload_descriptor.getNestedType(0);
-    internal_static_talon_data_FileUpload_RequiredHeadersEntry_fieldAccessorTable = new
-      com.google.protobuf.GeneratedMessage.FieldAccessorTable(
-        internal_static_talon_data_FileUpload_RequiredHeadersEntry_descriptor,
-        new java.lang.String[] { "Key", "Value", });
     internal_static_talon_data_HandleGrant_descriptor =
-      getDescriptor().getMessageType(3);
+      getDescriptor().getMessageType(2);
     internal_static_talon_data_HandleGrant_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_HandleGrant_descriptor,
         new java.lang.String[] { "Id", "Namespace", "Kind", "TargetId", "Agent", "SessionId", "Operations", "AudienceAgent", "AudienceSessionId", "ExpiresAt", "CreatedAt", });
     internal_static_talon_data_Principal_descriptor =
-      getDescriptor().getMessageType(4);
+      getDescriptor().getMessageType(3);
     internal_static_talon_data_Principal_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_Principal_descriptor,
@@ -26825,13 +24108,13 @@ java.lang.String defaultValue) {
         internal_static_talon_data_Principal_MetadataEntry_descriptor,
         new java.lang.String[] { "Key", "Value", });
     internal_static_talon_data_SessionMessagePart_descriptor =
-      getDescriptor().getMessageType(5);
+      getDescriptor().getMessageType(4);
     internal_static_talon_data_SessionMessagePart_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_SessionMessagePart_descriptor,
         new java.lang.String[] { "Id", "PartType", "Content", "Name", "PayloadJson", "CreatedAt", "Object", });
     internal_static_talon_data_SessionMessage_descriptor =
-      getDescriptor().getMessageType(6);
+      getDescriptor().getMessageType(5);
     internal_static_talon_data_SessionMessage_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_SessionMessage_descriptor,
@@ -26843,7 +24126,7 @@ java.lang.String defaultValue) {
         internal_static_talon_data_SessionMessage_LabelsEntry_descriptor,
         new java.lang.String[] { "Key", "Value", });
     internal_static_talon_data_Session_descriptor =
-      getDescriptor().getMessageType(7);
+      getDescriptor().getMessageType(6);
     internal_static_talon_data_Session_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_Session_descriptor,
@@ -26861,7 +24144,7 @@ java.lang.String defaultValue) {
         internal_static_talon_data_Session_LabelsEntry_descriptor,
         new java.lang.String[] { "Key", "Value", });
     internal_static_talon_data_ChannelMessage_descriptor =
-      getDescriptor().getMessageType(8);
+      getDescriptor().getMessageType(7);
     internal_static_talon_data_ChannelMessage_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_ChannelMessage_descriptor,
@@ -26873,19 +24156,19 @@ java.lang.String defaultValue) {
         internal_static_talon_data_ChannelMessage_LabelsEntry_descriptor,
         new java.lang.String[] { "Key", "Value", });
     internal_static_talon_data_Knowledge_descriptor =
-      getDescriptor().getMessageType(9);
+      getDescriptor().getMessageType(8);
     internal_static_talon_data_Knowledge_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_Knowledge_descriptor,
         new java.lang.String[] { "Path", "Content", "UpdatedAt", "Namespace", "Name", });
     internal_static_talon_data_KnowledgeSearchResult_descriptor =
-      getDescriptor().getMessageType(10);
+      getDescriptor().getMessageType(9);
     internal_static_talon_data_KnowledgeSearchResult_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_KnowledgeSearchResult_descriptor,
         new java.lang.String[] { "Path", "Snippet", "Score", "Timestamp", "Namespace", });
     internal_static_talon_data_WorkflowRun_descriptor =
-      getDescriptor().getMessageType(11);
+      getDescriptor().getMessageType(10);
     internal_static_talon_data_WorkflowRun_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_WorkflowRun_descriptor,
@@ -26897,13 +24180,13 @@ java.lang.String defaultValue) {
         internal_static_talon_data_WorkflowRun_LabelsEntry_descriptor,
         new java.lang.String[] { "Key", "Value", });
     internal_static_talon_data_WorkflowStepRun_descriptor =
-      getDescriptor().getMessageType(12);
+      getDescriptor().getMessageType(11);
     internal_static_talon_data_WorkflowStepRun_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_WorkflowStepRun_descriptor,
         new java.lang.String[] { "Id", "StepId", "Attempt", "Status", "InputJson", "OutputJson", "Error", "ChildSessionId", "ChildWorkflowRunId", "ResumeJson", "SuspendJson", "CreatedAt", "UpdatedAt", "NextRetryAt", "TimeoutAt", "WaitWakeupHandle", "WaitUntilAt", });
     internal_static_talon_data_WorkflowRunEvent_descriptor =
-      getDescriptor().getMessageType(13);
+      getDescriptor().getMessageType(12);
     internal_static_talon_data_WorkflowRunEvent_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_talon_data_WorkflowRunEvent_descriptor,
