+++ sdk/js/talon-client/src/gen/proto/data/data_pb.ts
@@ -316,150 +316,6 @@ export class Artifact extends Message<Artifact> {
 }
 
 /**
- * Pending direct-to-object-storage upload for a namespace File.
- *
- * The target File is not committed to this object until CompleteFileUpload
- * validates the uploaded object metadata and patches File.status.objectRef.
- *
- * @generated from message talon.data.FileUpload
- */
-export class FileUpload extends Message<FileUpload> {
-  /**
-   * @generated from field: string id = 1;
-   */
-  id = "";
-
-  /**
-   * @generated from field: string namespace = 2;
-   */
-  namespace = "";
-
-  /**
-   * @generated from field: string file_name = 3;
-   */
-  fileName = "";
-
-  /**
-   * @generated from field: string file_uid = 4;
-   */
-  fileUid = "";
-
-  /**
-   * @generated from field: string path = 5;
-   */
-  path = "";
-
-  /**
-   * @generated from field: string media_type = 6;
-   */
-  mediaType = "";
-
-  /**
-   * @generated from field: int32 purpose = 7;
-   */
-  purpose = 0;
-
-  /**
-   * @generated from field: int32 index_policy = 8;
-   */
-  indexPolicy = 0;
-
-  /**
-   * @generated from field: int32 retention = 9;
-   */
-  retention = 0;
-
-  /**
-   * @generated from field: string object_key = 10;
-   */
-  objectKey = "";
-
-  /**
-   * @generated from field: uint64 expected_size_bytes = 11;
-   */
-  expectedSizeBytes = protoInt64.zero;
-
-  /**
-   * @generated from field: string expected_sha256 = 12;
-   */
-  expectedSha256 = "";
-
-  /**
-   * @generated from field: map<string, string> required_headers = 13;
-   */
-  requiredHeaders: { [key: string]: string } = {};
-
-  /**
-   * @generated from field: string created_by_agent = 14;
-   */
-  createdByAgent = "";
-
-  /**
-   * @generated from field: string created_by_session_id = 15;
-   */
-  createdBySessionId = "";
-
-  /**
-   * @generated from field: int64 expires_at = 16;
-   */
-  expiresAt = protoInt64.zero;
-
-  /**
-   * @generated from field: int64 created_at = 17;
-   */
-  createdAt = protoInt64.zero;
-
-  /**
-   * @generated from field: bool consumed = 18;
-   */
-  consumed = false;
-
-  constructor(data?: PartialMessage<FileUpload>) {
-    super();
-    proto3.util.initPartial(data, this);
-  }
-
-  static readonly runtime: typeof proto3 = proto3;
-  static readonly typeName = "talon.data.FileUpload";
-  static readonly fields: FieldList = proto3.util.newFieldList(() => [
-    { no: 1, name: "id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 2, name: "namespace", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 3, name: "file_name", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 4, name: "file_uid", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 5, name: "path", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 6, name: "media_type", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 7, name: "purpose", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
-    { no: 8, name: "index_policy", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
-    { no: 9, name: "retention", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
-    { no: 10, name: "object_key", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 11, name: "expected_size_bytes", kind: "scalar", T: 4 /* ScalarType.UINT64 */ },
-    { no: 12, name: "expected_sha256", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 13, name: "required_headers", kind: "map", K: 9 /* ScalarType.STRING */, V: {kind: "scalar", T: 9 /* ScalarType.STRING */} },
-    { no: 14, name: "created_by_agent", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 15, name: "created_by_session_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
-    { no: 16, name: "expires_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
-    { no: 17, name: "created_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
-    { no: 18, name: "consumed", kind: "scalar", T: 8 /* ScalarType.BOOL */ },
-  ]);
-
-  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): FileUpload {
-    return new FileUpload().fromBinary(bytes, options);
-  }
-
-  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): FileUpload {
-    return new FileUpload().fromJson(jsonValue, options);
-  }
-
-  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): FileUpload {
-    return new FileUpload().fromJsonString(jsonString, options);
-  }
-
-  static equals(a: FileUpload | PlainMessage<FileUpload> | undefined, b: FileUpload | PlainMessage<FileUpload> | undefined): boolean {
-    return proto3.util.equals(FileUpload, a, b);
-  }
-}
-
-/**
  * Opaque access grant for a File or Artifact handle.
  *
  * Handle strings resolve to these KV-backed grant records. Callers must present
